Implemented Stamina Regeneration system

+++ Assets/Scripts/StaminaBar.cs
@@ -9,13 +9,13 @@ namespace CH
     {
         public Slider slider;
 
-        public void SetMaxStamina(int maxStamina)
+        public void SetMaxStamina(float maxStamina)
         {
             slider.maxValue = maxStamina;
             slider.value = maxStamina;
         }
 
-        public void SetCurrentStamina(int currentStamina)
+        public void SetCurrentStamina(float currentStamina)
         {
             slider.value = currentStamina;
         }
